Add option to force restart worker in any case
Add option to force restart worker even if there was no configuration
change which would trigger the worker restart. This option is useful for
playbooks which use this role and need to force the worker restart based
on external events (such as installing a new CA certificate on the
system).

Signed-off-by: Tomáš Hozza <thozza@redhat.com>