Ensure that worker is not restarted if not needed
Previously when running the role on a fresh system, it would configure
the worker and notify the handler to restart it and at the end of
configuration it would start the worker service. As a result, the worker
would get started and then immediately restarted by the handler.

Modify the role to trigger handlers before ensuring that the worker
service is started and enabled. This ensures that the worker is not
unnecessarily restarted.

Modify the default test to check that the worker was not restarted.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>